Issue #34 removed interface from eventservice since it is no longer a managed bean
